move filters to common folder
